+++ docs/changelog.md
@@ -7,6 +7,8 @@
 - `borax.calendars.festivals2`
   - 新增 `Festival.gets` 获取属性方法
   - 新增 `Festival.description` 属性，表示规范化中文描述
+- `borax.calendars.utils`
+  - 新增 `ThreeNineUtils` 三伏九九天解析工具
 
 ## v3.5.0 (20211115)